.history/requirements_20230411163424.txt
click==8.1.3
configobj==5.0.8
Flask==2.2.3
greenlet==2.0.2
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
lz4==4.3.2
MarkupSafe==2.1.2
PyMySQL==1.0.3
python-dateutil==2.8.2
PyYAML==6.0
ruamel.yaml==0.17.21
ruamel.yaml.base==0.3.2
ruamel.yaml.clib==0.2.7
ruamel.yaml.cmd==0.6.3
ruamel.yaml.convert==0.3.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
